Simplify autocompletion code and allow options almost anywhere.
Massive simplification of generated Bash auto completion code. The old
code did a lot of fancy validation which is in reality not necessary to
provide a good command line user experience. The code was complex, large
and made adding new functionality to the command line interface a
laborious task.

Also, to keep in line with changes to command line parsing, allow
completion of options almost anywhere.